About - VSR Park Street
A project spread in 2.85 acres in Sector 85 Gurugram, Strategically located on 157 meters wide road, which is just 1 km from Dwarka Expressway and is next to commercial Sector 88, which is the future market for enterprise. Changing with the changing times, this unique project located This new corridor in proximity with the Pataudi Road has earned prominence due to its accessibility from National Highway - 8 (NH-8) and Dwarka expressway and easy reach to Industrial Model Township (IMT) Manesar.